型のすり合わせ

+++ domains/src/value_objects/timestamp.rs
@@ -4,12 +4,12 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct TimeStamp {
-    pub seconds: i64,
+    pub seconds: u64,
     pub description: TimeStampDescription,
 }
 
 impl TimeStamp {
-    pub fn new(seconds: i64, description: TimeStampDescription) -> Self {
+    pub fn new(seconds: u64, description: TimeStampDescription) -> Self {
         TimeStamp {
             seconds,
             description,
@@ -17,7 +17,7 @@ impl TimeStamp {
     }
 
     pub fn to_datetime(&self) -> DateTime<Utc> {
-        Utc.timestamp_opt(self.seconds, 0)
+        Utc.timestamp_opt(self.seconds as i64, 0)
             .single()
             .expect("Invalid timestamp")
     }
